// src/config/database/mysql/enums/mysql.environments.enum.ts
export enum MySQLEnvironments{
    HOST = "mysql.HOST",
    PORT = "mysql.PORT",
    USERNAME = "mysql.USERNAME",
    PASSWORD = "mysql.PASSWORD",
    DATABASE_NAME = "mysql.DATABASE_NAME", 
    ENTITIES = "mysql.ENTITIES",
    AUTO_LOAD_ENTITIES = "mysql.AUTO_LOAD_ENTITIES",
    MIGRATIONS = "mysql.MIGRATIONS",
    MIGRATIONS_TABLE_NAME = "mysql.MIGRATIONS_TABLE_NAME",
    MIGRATIONS_RUN = "mysql.MIGRATIONS_RUN",
    LOGGING = "mysql.LOGGING",
    SYNCHRONIZE = "mysql.SYNCHRONIZE",
    MIGRATIONS_DIR = "mysql.MIGRATIONS_DIR"
}